+++ packages/shards-process-manager-core/src/collections.ts
@@ -1,3 +1,0 @@
-import { ProcessRecord } from './types';
-
-export const ProcessCollection = new Mongo.Collection<ProcessRecord>('background-process-manager---processes');